# src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  title: Vendor Service
  description: Vendor Service CRUD and serch API.
  version: 1.0.0
tags:
  - name: CRUD Operations
    description: Manage individual vendor profiles
  - name: Search Operations
    description: Retrieve a list of vendors
paths:
  /vendor:
    post:
      tags:
        - CRUD Operations
      summary: Add a vendor to the system
      operationId: addVendor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VendorCreateDto'
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VendorDto'
        406:
          description: >
            Something wrong with the vendor data submitted, list of error fields returned with details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorList'
              example:
                - code: name
                  message: Value is too long
        default:
          description: An error occured while saving, when it's an error with fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: 400
                message: Invalid vendor object
  /vendor/{id}:
    get:
      tags:
        - CRUD Operations
      summary: Get vendor details by id
      operationId: getVendorById
      parameters:
        - name: id
          in: path
          description: The id of the vendor to return
          required: true
          schema:
            type: string
      responses:
        200:
          description: Returns vendor details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VendorDto'
        404:
          description: Vendor not found
    delete:
      tags:
        - CRUD Operations
      summary: Delete Vendor
      operationId: deleteVendorById
      parameters:
        - name: id
          in: path
          description: The id of the vendor to delete
          required: true
          schema:
            type: string
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VendorDto'
        404:
          description: Vendor not found
    post:
      tags:
        - CRUD Operations
      summary: update Vendor
      operationId: updateVendorById
      parameters:
        - name: id
          description: The id of the vendor to update
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VendorDto'
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VendorDto'
        406:
          description: >
            Something wrong with the vendor data submitted, list of error fields returned with details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorList'
              example:
                - code: lastName
                  message: Value is too long
                - code: phin
                  message: Invalid number format
        default:
          description: An error occured while saving, when it's an error with fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: 400
                message: Invalid vendor object

components:
  schemas:
    ## This is a pseudo hack, because the generator doesn't like
    ## having "unnamed" schemas (should just go in the request definition
    ## for the update method
    ## https://github.com/OpenAPITools/openapi-generator/issues/2892
    VendorCreateDto:
      description: VendorDTO object plus information on required fields
      type: object
      allOf:
        - $ref: '#/components/schemas/VendorDto'
        - required:
            - name
            - contactDetails
        ## although correct syntax (you should be able to put at schema level too - separate bug)
        ## the generator doesn't merge properties either way you do it (meaning we should only
        ## have minItems and nullable listed below, but to work around the generator bug...
        - properties:
            contactDetails:
              type: array
              minItems: 1
              maxItems: 8
              nullable: false
              uniqueItems: true
              items:
                $ref: "#/components/schemas/ContactDetailDto"
    VendorDto:
      type: object
      properties:
        id:
          type: string
          description: The UUID for the vendor across all systems
          readOnly: true
        name:
          type: string
          example: All Widgets Inc.
          maxLength: 120
          description: The vendor name
        description:
          type: string
          example: Sells widgets and stuff to us.
          maxLength: 1000
          description: Some details about the vendor
        contactDetails:
          type: array
          uniqueItems: true
          maxItems: 8
          items:
            $ref: "#/components/schemas/ContactDetailDto"
    ContactDetailDto:
      type: object
      properties:
        detailType:
          type: string
          enum: [ "Office", "Representative", "Alternative" ]
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
    ErrorList:
      type: array
      items:
        $ref: "#/components/schemas/Error"
    Error:
      type: object
      properties:
        code:
          type: string
          description: A code specific to the particular message
        message:
          type: string
          description: The related error message